refactor: split unit and end-to-end tests #166
